Codebase verified!
Testing and conferencing finally revealed that the last major source of
discrepancy between my code and Byeonghee's code was our definition of
the two-sided discrete derivative.  Though we had agreed on our fiducial
parameters theta and step sizes dtheta, By was dividing dtheta by 2 in
the numerator while I was multiplying dtheta by two in the denominator,
effectively making my step sizes twice what his were.  This raises the
additional question of whether or not we have converged on the correct
integrals given the sensitivity to step sizes, but we are tabling this
problem for now as it is expected to have less of an affect on our
results than the noise contributions which I still have yet to
incorporate.