Noise Incorporated
Primary CMB TT, TE, and EE now inculded in observables. These are
assumed to have no cross correlation with the kappa and galaxy
observables and thus are implemented as a separate Fisher matrix which
can be added to the k,g matrix.

The lensing reconstruction noise, T and E detector noise, and galaxy
shot noise have all been implemented and noise levels set to CMB-S4 and
LSST expectations.